pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["openanomaly"]

[project]
name = "openanomaly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.6.1",
    "fastapi>=0.128.0",
    "pymongo>=4.15.5",
    "motor>=3.3.2",
    "pyyaml>=6.0",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.6.0",
    "opentelemetry-exporter-prometheus-remote-write>=0.43b0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-api>=1.22.0",
    "redis>=5.0.0",
    "django>=5.0",
    "django-mongodb-backend>=5.0",
    "celery-redbeat>=2.0.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
    "pytest-asyncio>=1.3.0",
    "django-environ>=0.11.2",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
ml = [
    "torch>=2.0.0",
    "chronos-forecasting>=2.2.2",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
